rename package to common
