+++ acceptance_test.go
@@ -18,7 +18,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/conduitio/conduit-connector-kafka/config"
+	"github.com/conduitio/conduit-connector-kafka/common"
 	"github.com/conduitio/conduit-connector-kafka/source"
 	"github.com/conduitio/conduit-connector-kafka/test"
 	sdk "github.com/conduitio/conduit-connector-sdk"
@@ -60,7 +60,7 @@ type AcceptanceTestDriver struct {
 // ReadFromDestination is overwritten because the source connector uses a consumer
 // group which results in slow reads. This speeds up the destination tests.
 func (d AcceptanceTestDriver) ReadFromDestination(t *testing.T, records []sdk.Record) []sdk.Record {
-	cfg := test.ParseConfigMap[config.Config](t, d.Config.SourceConfig)
+	cfg := test.ParseConfigMap[common.Config](t, d.Config.SourceConfig)
 	kgoRecs := test.Consume(t, cfg, len(records))
 
 	recs := make([]sdk.Record, len(kgoRecs))
